requirements.txt
Django==1.8.5
psycopg2==2.6.2
djangorestframework==3.3.0
pytz==2016.10
python-dotenv==0.6.1
requests==2.12.4

pytest==3.0.5
pytest-django==3.1.2
model-mommy==1.3.1
freezegun==0.3.8
httpretty==0.8.14
mock==2.0.0
